feat: update text renderer
