+++ packages/presets/src/ai/utils.ts
@@ -1,12 +1,7 @@
-import type { EditorHost } from '@blocksuite/block-std';
 import {
   type AffineAIPanelWidget,
   type AffineAIPanelWidgetConfig,
-  defaultImageProxyMiddleware,
-  MarkdownAdapter,
 } from '@blocksuite/blocks';
-import type { Doc } from '@blocksuite/store';
-import { DocCollection, Job } from '@blocksuite/store';
 
 import type { CopilotClient } from './copilot-client.js';
 
@@ -55,25 +50,4 @@ export function getGenerateAnswer({
       .catch(console.error);
   };
   return generateAnswer;
-}
-
-export async function markDownToDoc(host: EditorHost, answer: string) {
-  const schema = host.std.doc.collection.schema;
-  // Should not create a new doc in the original collection
-  const collection = new DocCollection({ schema });
-  const job = new Job({
-    collection,
-    middlewares: [defaultImageProxyMiddleware],
-  });
-  const mdAdapter = new MarkdownAdapter();
-  mdAdapter.applyConfigs(job.adapterConfigs);
-  const snapshot = await mdAdapter.toDocSnapshot({
-    file: answer,
-    assets: job.assetsManager,
-  });
-  const doc = await job.snapshotToDoc(snapshot);
-  if (!doc) {
-    console.error('Failed to convert markdown to doc');
-  }
-  return doc as Doc;
 }
