+++ packages/presets/src/ai/utils/markdown-utils.ts
@@ -1,8 +1,11 @@
 import type { EditorHost } from '@blocksuite/block-std';
-import { MarkdownAdapter } from '@blocksuite/blocks';
+import {
+  defaultImageProxyMiddleware,
+  MarkdownAdapter,
+} from '@blocksuite/blocks';
 import { assertExists } from '@blocksuite/global/utils';
-import type { BlockModel } from '@blocksuite/store';
-import { Job, type Slice } from '@blocksuite/store';
+import type { BlockModel, Doc } from '@blocksuite/store';
+import { DocCollection, Job, type Slice } from '@blocksuite/store';
 
 export async function getMarkdownFromSlice(host: EditorHost, slice: Slice) {
   const job = new Job({ collection: host.std.doc.collection });
@@ -69,4 +72,25 @@ export async function insertFromMarkdown(
   }
 
   return models;
+}
+
+export async function markDownToDoc(host: EditorHost, answer: string) {
+  const schema = host.std.doc.collection.schema;
+  // Should not create a new doc in the original collection
+  const collection = new DocCollection({ schema });
+  const job = new Job({
+    collection,
+    middlewares: [defaultImageProxyMiddleware],
+  });
+  const mdAdapter = new MarkdownAdapter();
+  mdAdapter.applyConfigs(job.adapterConfigs);
+  const snapshot = await mdAdapter.toDocSnapshot({
+    file: answer,
+    assets: job.assetsManager,
+  });
+  const doc = await job.snapshotToDoc(snapshot);
+  if (!doc) {
+    console.error('Failed to convert markdown to doc');
+  }
+  return doc as Doc;
 }
